Unmount docs again
